ControlCmdString neu: Anzeige in hex Ersatz NSMomentaryLight > 	[NSButtonTypeMomentaryPushIn
rHomeData: NSURLSession neu
rTagPlanBalken permanent neu

+++ rAVR_DS.h
@@ -18,7 +18,7 @@ NSMutableArray* WochenplanTabelle;
 - (void)setWochenplan:(NSArray*) derWochenplan;
 - (void)clearWochenplan;
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView;
+- (long)numberOfRowsInTableView:(NSTableView *)aTableView;
 - (id)tableView:(NSTableView *)aTableView
     objectValueForTableColumn:(NSTableColumn *)aTableColumn 
 			row:(int)rowIndex;
