+++ rWEBDATA_DS.m
@@ -86,7 +86,7 @@
 
 
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (long)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     return [ValueKeyTabelle count];
 }
